change style and rename file names
