+++ src/webportal/src/plugins/job-submission/components/completionSection.jsx
@@ -24,40 +24,49 @@
  */
 
 import React from 'react';
-import {TextField} from 'office-ui-fabric-react';
-import {BasicSection} from './BasicSection';
+import {Stack} from 'office-ui-fabric-react';
 import PropTypes from 'prop-types';
-import {Deployment} from '../models/deployment';
+import {BasicSection} from './basicSection';
+import {Completion} from '../models/completion';
+import {CSpinButton} from './customizedComponents';
+import {getCompletionSectionStyle} from './formStyle';
+import {FormShortSection} from './formPage';
 
-export const DeploymentSection = (props) => {
+const completionSectionStyle = getCompletionSectionStyle();
+
+export const CompletionSection= (props) => {
   const {onChange, value} = props;
-  const {preCommands, postCommands} = value;
+  const {minFailedInstances, minSuceedInstances} = value;
 
   const _onChange = (keyName, newValue) => {
-    const updatedDeployment = new Deployment(value);
-    updatedDeployment[keyName] = newValue;
+    const completion = new Completion(value);
+    completion[keyName] = newValue;
     if (onChange !== undefined) {
-      onChange(updatedDeployment);
+      onChange(completion);
     }
   };
 
   return (
-    <BasicSection sectionLabel={'Deployment'} sectionOptional>
-      <TextField label={'PreCommands'}
-                 value={preCommands}
-                 onBlur={(e) => _onChange('preCommands', e.target.value)}
-                 multiline
-                 rows={3}/>
-      <TextField label={'PostCommands'}
-                 value={postCommands}
-                 multiline
-                 rows={3}
-                 onBlur={(e) => _onChange('postCommands', e.target.value)}/>
+    <BasicSection sectionLabel={'Completion'} sectionOptional>
+      <FormShortSection gap='m'>
+        <Stack horizontal gap='s1'>
+          <CSpinButton label={'Min Failed Instances'}
+                       value={minFailedInstances}
+                       styles={completionSectionStyle.spinButton}
+                       onChange={(v) => _onChange('minFailedInstances', v)}/>
+        </Stack>
+        <Stack horizontal gap='s1'>
+          <CSpinButton label={'Min Succeed Instances'}
+                       value={minSuceedInstances}
+                       styles={completionSectionStyle.spinButton}
+                       onChange={(v) => _onChange('minSuceedInstances', v)}/>
+        </Stack>
+      </FormShortSection>
     </BasicSection>
   );
 };
 
-DeploymentSection.propTypes = {
-  value: PropTypes.instanceOf(Deployment).isRequired,
+CompletionSection.propTypes = {
+  value: PropTypes.instanceOf(Completion).isRequired,
   onChange: PropTypes.func,
 };
