+++ src/webportal/src/plugins/job-submission/components/submissionSection.jsx
